site reload even when no zoomed video is generated

+++ zooming_app.py
@@ -8,6 +8,9 @@ import time
 app = Flask(__name__)
 
 
+global cropped_frames_path
+cropped_frames_path = ''
+
 #rendering the home page
 @app.route("/")
 def home():
@@ -54,7 +57,6 @@ def process():
     # saving output video with timestamp to avoid browser caching
     unique_suffix = str(time.time())
     output_video_name = f"cropped_video_output{unique_suffix}"
-    global cropped_frames_path
     # setting the path of the saved video
     cropped_frames_path = (
         f"static/imgs/{output_video_name}.mp4"
@@ -76,6 +78,8 @@ def process():
 # deleting the created video after the user presses the refresh button
 @app.route("/delete-file", methods=["POST"])
 def delete_file():
+    if not os.path.exists(cropped_frames_path):
+        return 'No file to delete'
     # removing the file when the user clicks on the refresh button
     try:
         os.remove(cropped_frames_path)
